refactor: extract openFileInEditor to shared utility
- Create src/common/vscodeEditor.ts with openFileInEditor and ensureFileOpen
- Update LspTools.ts to use shared utility, remove ~65 lines
- Update linter.ts to use ensureFileOpen, remove ~60 lines
- Remove debug logs from VscodeIntegration.ts, now 48 lines total
- Simplify getDiagnostics fallback logic
